Adding pattern matching for lists in scala
Need to add the same for Kotlin (which doesnt support what scala does), but can be simulated. This is possibly easiest to do in JS with the spread operator tbh.
Speaking of which, Kotlin can spread too, but not sure if it can do a destructure spread like Ruby/JS